backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
paho-mqtt~=2.1.0
pandas~=2.2.3
scipy==1.14.1